// gantryweb/Frontend/Gantry/src/app/common/models/data-feed/sport-bet-models.ts
export class SportBookEvent implements IChannelsObject {
    eventKey: number | null | undefined;
    eventName: string | null | undefined;
    eventStatus: string | null | undefined;
    displayStatus: string | null | undefined;
    displayOrder: number | null | undefined;
    hasBIRMarkets: string | null | undefined;
    eventSort: string | null | undefined;
    eventDateTime: Date | null | undefined;
    isEventStarted: boolean | null | undefined;
    isEventFinished: boolean | null | undefined;
    isEventResulted: boolean | null | undefined;
    isCashoutAvailable: boolean | null | undefined;
    channels: string[] | null | undefined;
    flags: string[] | null | undefined;
    meta: DataMeta | null | undefined;
    raceStage: string | null | undefined;
    offTime: Date | null | undefined;
    categoryName?: string | null | undefined;
    typeName?: string | null | undefined;
    typeFlagCode?: string | null | undefined;
    runnerCount?: number | null | undefined;
    eventTimePlusTypeName?: string | null | undefined;
    typeKey?: string | null | undefined;
}
// ...
export class SportBookResult {
    events: Map<number, SportBookEventStructured> = new Map<number, SportBookEventStructured>();
}

export class SportBookEventStructured extends SportBookEvent {
    markets: Map<number, SportBookMarketStructured> = new Map<number, SportBookMarketStructured>();
}

export class SportBookMarketStructured extends SportBookMarket {
    selections: Map<number, SportBookSelection> = new Map<number, SportBookSelection>();
}